requirements.txt
mysql-connector-python==8.3.0
pyaes==1.6.1
pyasn1==0.5.1
python-dotenv==1.0.1
rsa==4.9
Telethon==1.34.0